refactor: duplicated code removed
