+++ src/api/fetchersFunction/productMetaDataFetchers.ts
@@ -8,38 +8,9 @@ import {
   CTP_PROJECT_KEY,
   LS_KEY,
 } from '../../constants';
-import { getProductDetails } from '../graphql/productDetails';
-import apiRoot from '../apiRoot';
 import OpenAI from 'openai';
 import { getAllSavedRulesFromCtObj } from './ruleFetchers';
-
-export const getProductById = async (productId: string, locale?: string) => {
-  try {
-    const response = await apiRoot
-      .graphql()
-      .post({
-        body: {
-          query: getProductDetails(),
-          variables: {
-            id: productId,
-            Locale: locale,
-          },
-        },
-      })
-      .execute();
-
-    const product = response.body.data.product;
-
-    if (!product) {
-      return `Product with ID ${productId} not found.`;
-    }
-    return product;
-  } catch (error) {
-    console.error(`Error retrieving product by ID ${productId}:`, error);
-
-    return 'Failed to retrieve product details';
-  }
-};
+import { getProductById } from './utils';
 
 export const generateProductMetaData = async (
   productId: string,
