+++ src/components/Renderers/ActionRendererSEO.tsx
@@ -1,4 +1,3 @@
-import PrimaryButton from '@commercetools-uikit/primary-button';
 import { useAppContext } from '../../context/AppContext';
 import { useApplicationContext } from '@commercetools-frontend/application-shell-connectors';
 import {
@@ -6,6 +5,7 @@ import {
   updateProductSeoMeta,
 } from '../../api/fetchersFunction/seoMetaDataFetchers';
 import { seoMatchData } from '../../api/fetchersFunction/utils';
+import ActionRenderButtons from './ActionRenderButtons';
 
 export default (props: any) => {
   const { setState } = useAppContext();
@@ -84,23 +84,10 @@ export default (props: any) => {
   };
 
   return (
-    <div style={{ display: 'flex' }}>
-      <div>
-        <PrimaryButton
-          size="medium"
-          label="Generate"
-          onClick={() => handleGenerateClick(props)}
-          isDisabled={false}
-        />
-      </div>
-      <div style={{ marginInline: '6px' }}>
-        <PrimaryButton
-          size="medium"
-          label="Apply"
-          onClick={() => handleApplyClick(props.rowIndex)}
-          isDisabled={false}
-        />
-      </div>
-    </div>
+    <ActionRenderButtons
+    handleGenerate={handleGenerateClick}
+    handleApply={handleApplyClick}
+    allProps={props}
+  />
   );
 };
